Refactor Draft/TodoItems into separate files
