+++ web-js/src/components/DraftTodoItem.js
@@ -1,0 +1,57 @@
+import React from "react";
+import {
+  TextField,
+  Button,
+  IconButton,
+  ListItem,
+  ListItemSecondaryAction,
+  ListItemText,
+  withStyles,
+} from "@material-ui/core";
+import ClearIcon from "@material-ui/icons/Clear";
+
+const ListItemWithTwoSecondaryActions = withStyles({
+  secondaryAction: {
+    paddingRight: "120px",
+  },
+})(ListItem);
+
+export function DraftTodoItem({ todo, todoActions, draftTodoActions }) {
+  return (
+    <ListItemWithTwoSecondaryActions>
+      <ListItemText inset>
+        <TextField
+          style={{ width: "100%" }}
+          placeholder="What needs doing?"
+          size="small"
+          value={todo.summary}
+          onChange={(e) => {
+            draftTodoActions.setDraftTodoSummary(todo, e.target.value);
+          }}
+        />
+      </ListItemText>
+      <ListItemSecondaryAction>
+        <Button
+          variant="outlined"
+          color="secondary"
+          size="small"
+          onClick={() => {
+            todoActions.saveTodo(todo)
+            draftTodoActions.deleteDraftTodo(todo)
+          }}
+        >
+          Save
+        </Button>
+        <IconButton
+          edge="end"
+          size="small"
+          onClick={() => {
+            draftTodoActions.deleteDraftTodo(todo)
+          }}
+        >
+          <ClearIcon />
+        </IconButton>
+      </ListItemSecondaryAction>
+    </ListItemWithTwoSecondaryActions>
+  );
+}
