+++ web-js/src/components/TodoItemsPage.js
@@ -1,21 +1,14 @@
 import React from "react";
 import {
   Container,
-  TextField,
   Button,
-  IconButton,
   Typography,
-  Checkbox,
   List,
-  ListItem,
-  ListItemIcon,
-  ListItemSecondaryAction,
-  ListItemText,
-  withStyles,
   LinearProgress,
 } from "@material-ui/core";
 import AddIcon from "@material-ui/icons/Add";
-import ClearIcon from "@material-ui/icons/Clear";
+import { TodoItem } from './TodoItem'
+import { DraftTodoItem } from './DraftTodoItem'
 import { useTodos } from "./useTodos";
 import { useDraftTodos } from "./useDraftTodos";
 import { useShowLoader } from "./util-hooks";
@@ -64,76 +57,5 @@ export function TodoItemsPage() {
       <MoreInfoGraphiQL />
       <MoreInfoTemplateAndDocs />
     </Container>
-  );
-}
-
-function TodoItem({ isDraft = false, todo, todoActions }) {
-  const toggleIsCompleted = () => {
-    todoActions.toggleTodo(todo);
-  };
-  return (
-    <ListItem>
-      <ListItemIcon>
-        <Checkbox
-          edge="start"
-          checked={todo.isComplete}
-          onClick={toggleIsCompleted}
-        />
-      </ListItemIcon>
-      <ListItemText>{todo.summary}</ListItemText>
-      <ListItemSecondaryAction>
-        <IconButton edge="end" size="small" 
-          onClick={() => {
-            todoActions.deleteTodo(todo)
-          }}>
-          <ClearIcon />
-        </IconButton>
-      </ListItemSecondaryAction>
-    </ListItem>
-  );
-}
-
-const ListItemWithTwoSecondaryActions = withStyles({
-  secondaryAction: {
-    paddingRight: "120px",
-  },
-})(ListItem);
-function DraftTodoItem({ todo, todoActions, draftTodoActions }) {
-  return (
-    <ListItemWithTwoSecondaryActions>
-      <ListItemText inset>
-        <TextField
-          style={{ width: "100%" }}
-          placeholder="What needs doing?"
-          size="small"
-          value={todo.summary}
-          onChange={(e) => {
-            draftTodoActions.setDraftTodoSummary(todo, e.target.value);
-          }}
-        />
-      </ListItemText>
-      <ListItemSecondaryAction>
-        <Button
-          variant="outlined"
-          color="secondary"
-          size="small"
-          onClick={() => {
-            todoActions.saveTodo(todo)
-            draftTodoActions.deleteDraftTodo(todo)
-          }}
-        >
-          Save
-        </Button>
-        <IconButton
-          edge="end"
-          size="small"
-          onClick={() => {
-            draftTodoActions.deleteDraftTodo(todo)
-          }}
-        >
-          <ClearIcon />
-        </IconButton>
-      </ListItemSecondaryAction>
-    </ListItemWithTwoSecondaryActions>
   );
 }
